requirement.txt
appdirs==1.4.4
asgiref==3.6.0
attrs==22.2.0
black==22.12.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colorlog==4.8.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
exceptiongroup==1.1.0
filetype==1.2.0
har2case==0.3.1
httprunner==4.3.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.9.5
jsonpath==0.82
jsonschema==3.2.0
loguru==0.4.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
pluggy==0.13.1
py==1.11.0
pydantic==1.8.2
PyJWT==1.7.1
PyMySQL==1.0.3
pyrsistent==0.19.3
pytest==7.2.2
pytest-metadata==2.0.4
pytest-testreport==1.1.6
pytz==2022.7.1
PyYAML==5.4.1
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.9.1
rest-framework-simplejwt==0.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==0.14.4
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
